requirements.txt
# ============================================================
# quant_etl — 固定版本依賴清單
#
# 產出方式：依據 Python 3.13 實際執行環境掃描，
# 只列入專案程式碼直接 import 的套件。
#
# 安裝方式：
#   pip install -r requirements.txt
# ============================================================

# ── 資料處理 ───────────────────────────────────────────────
pandas==2.3.3
numpy==2.4.3

# ── 市場資料 ───────────────────────────────────────────────
yfinance==1.2.0

# ── 總經資料（FRED）────────────────────────────────────────
fredapi==0.5.2

# ── 資料庫（PostgreSQL）────────────────────────────────────
psycopg2-binary==2.9.11

# ── 環境設定 ───────────────────────────────────────────────
python-dotenv==1.2.2

# ── 報告 / Email ───────────────────────────────────────────
Markdown==3.10.2

# HTTP (LINE Messaging API)
requests==2.32.3

# Webhook Server（關鍵字查詢模式）
flask==3.0.3
